+++ lib/util/random.h
@@ -3,6 +3,7 @@
 
 #pragma once
 
+#include <stddef.h>
 #include <stdint.h>
 #include <util/api.h>
 
@@ -17,7 +18,7 @@ extern "C" {
 ///
 /// @param bound Exclusive upper bound on the sequence
 /// @return A permutation of `[0, bound - 1]`
-UTIL_API int *gv_permutation(int bound);
+UTIL_API size_t *gv_permutation(size_t bound);
 
 /// generate a random number in the range `[0, bound - 1]`
 ///
@@ -26,7 +27,7 @@ UTIL_API int *gv_permutation(int bound);
 ///
 /// @param bound Exclusive upper bound on random number generation
 /// @return A random number drawn from a uniform distribution
-UTIL_API int gv_random(int bound);
+UTIL_API size_t gv_random(size_t bound);
 
 /// generate a random 64-bit unsigned number in the range `[0, bound - 1]`
 ///
